discover: Check if the kernel image has Ultravisor support
The PPC kernel image has an ELF Note 'namespace' called 'PowerPC'
to store capabilities and information which can be used by a
bootloader or userland. The capabilities can be accessed using
the 'type' PPC_ELFNOTE_CAPABILITIES which returns a bitmap
as 'descriptor' field.

Bit 0 in this bitmap indicates that the powerpc kernel binary
knows how to run in an ultravisor-enabled system. So, using this
bit, the petitboot can decide to abort the boot if the kernel is
incompatible, avoiding the crash later.

This validation only occours on PowerPC ultravisor-system and if
the config 'preboot check' in UI screen is enabled.

Signed-off-by: Maxiwell S. Garcia <maxiwell@linux.ibm.com>