discover/grub2: Use getopt for `search` argument parsing
The search command will be extended to add the full set of grub2-style
arguments, so switch to using getopt, rather than manual parsing.

This means we now support `--set=foo` and `--set foo` style arguments,
both of which appear in the docs and common grub configs.

Also, add a small test for the search argument handling.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>