discover/grub2: Add the '-e' test support
Grub2 allows a special-case of file test using the '-e' operator, where
the path can be empty, and the device existance is tested.
E.g.:
  if [ -e (md/md-boot) ]; then

Add the support for testing this condition, as well as some tests for
this scenario (also fix an existing testcase where this was wrongly
passing).

This fixes the following RH CoreOS bug:
  https://bugzilla.redhat.com/show_bug.cgi?id=1915540

Signed-off-by: Klaus Heinrich Kiwi <klaus@linux.vnet.ibm.com>
Signed-off-by: Jeremy Kerr <jk@ozlabs.org>