discover: Fix comment in user_event_handle_message()
The comment says that user_event_url() will steal the event context, but
that's not true. It actually allocates another event struct:

    user_event_url(uev, event)
        device_handler_process_url(handler, url, mac, ip)
            event = talloc_zero(handler, struct event);
            ...
            talloc_steal(ctx, event);

The variable names are the same, which probably led to the confusion,
but the "event" being stolen above is not the one from user_event_url().

This causes me to wonder whether device_handler_process_url() really
even needs to call talloc_steal() at all. There could be some subtlety
of the asynchronous job setup that I'm missing.

For now, just correct the comment.

Signed-off-by: Reza Arbab <arbab@linux.ibm.com>