lib/process: Handle SIGCHLD correctly
SIGCHLD is not guaranteed to be sent for all children, especially
if 2 children die at the same time.

Instead of trying to find the specific child that died, just handle
all of the async processes that have finished running.

This removes a hang observed when downloading an image over the
network. 5 wget processes are started, and some of them finished
close enough together that only a single SIGCHLD was delivered to
pb-discover.

Signed-off-by: Lincoln Ramsay <lincoln.ramsay@opengear.com>