pb-console: Ensure kernel log output is reduced
In a typical user-separated installation of petitboot, the root user
runs 'pb-console --getty=/sbin/agetty -- -a petituser'. With this usage,
pb-console acts as a wrapper for agetty to autologin petituser.

During that login, it is actually petituser's .profile which runs
'pb-console' (without --getty) to launch the UI.

The part of the pb-console script that reduces kernel log output is in
the codepath of the latter invocation. It fails because we aren't root.

  $ dmesg -n 1
  dmesg: klogctl: Operation not permitted

Move this line so it runs in either invocation of pb-console. For good
measure, add a check to make sure we're root.

Signed-off-by: Reza Arbab <arbab@linux.ibm.com>