discover/grub2: Ignore bare values in blscfg files
When pb-discover encounters a bare value in a blscfg file,
bls_process_pair() will be called with name=0x0, leading to a segfault
when parsing starts:

	if (streq(name, "title")) { ... }

Add a parser testcase for Fedora CoreOS, the real world config
triggering this bug (with the bare value "grub_users"). To fix, ignore
bare values as we do in other parsers.

Signed-off-by: Reza Arbab <arbab@linux.ibm.com>